Minor: Move ObjectStoreRegistry to datafusion_execution
