+++ datafusion/core/src/datasource/mod.rs
@@ -25,9 +25,11 @@ pub mod file_format;
 pub mod listing;
 pub mod listing_table_factory;
 pub mod memory;
-pub mod object_store;
 pub mod streaming;
 pub mod view;
+
+// backwards compatibility
+pub use datafusion_execution::object_store;
 
 use futures::Stream;
