+++ datafusion/core/src/execution/runtime_env.rs
@@ -26,9 +26,11 @@ use std::collections::HashMap;
 
 use crate::datasource::datasource::TableProviderFactory;
 use crate::datasource::listing_table_factory::ListingTableFactory;
-use crate::datasource::object_store::ObjectStoreRegistry;
-use crate::execution::memory_pool::{GreedyMemoryPool, MemoryPool, UnboundedMemoryPool};
 use datafusion_common::DataFusionError;
+use datafusion_execution::{
+    memory_pool::{GreedyMemoryPool, MemoryPool, UnboundedMemoryPool},
+    object_store::ObjectStoreRegistry,
+};
 use object_store::ObjectStore;
 use std::fmt::{Debug, Formatter};
 use std::path::PathBuf;
