+++ datafusion/execution/Cargo.toml
@@ -36,8 +36,11 @@ path = "src/lib.rs"
 [dependencies]
 datafusion-common = { path = "../common", version = "19.0.0" }
 datafusion-expr = { path = "../expr", version = "19.0.0" }
+dashmap = "5.4.0"
 hashbrown = { version = "0.13", features = ["raw"] }
 log = "^0.4"
+object_store = "0.5.4"
 parking_lot = "0.12"
 rand = "0.8"
 tempfile = "3"
+url = "2.2"
