+++ datafusion/execution/src/lib.rs
@@ -17,4 +17,5 @@
 
 pub mod disk_manager;
 pub mod memory_pool;
+pub mod object_store;
 pub mod registry;
